requirements.txt
numpy
Pillow
streamlit
tensorflow
matplotlib
